UIDs are "unsigned long"s
we should define a central uid type, probably

+++ imap/append.c
@@ -1,5 +1,5 @@
 /* append.c -- Routines for appending messages to a mailbox
- * $Id: append.c,v 1.94 2002/07/24 19:30:31 rjs3 Exp $
+ * $Id: append.c,v 1.95 2002/12/02 22:30:22 leg Exp $
  *
  * Copyright (c)1998, 2000 Carnegie Mellon University.  All rights reserved.
  *
@@ -333,7 +333,7 @@ int append_commit(struct appendstate *as,
 int append_abort(struct appendstate *as)
 {
     int r = 0;
-    int uid;
+    unsigned long uid;
 
     if (as->s == APPEND_DONE) return 0;
     as->s = APPEND_DONE;
